Fixed errors with TwoWayLever and organized files.
