+++ Resources/Prototypes/Entities/Structures/Wallmounts/switch.yml
@@ -97,11 +97,10 @@
     - type: UseDelay
       delay: 0.2 # prevent light-toggling auto-clickers.
     - type: Appearance
-      visuals:
-        - type: TwoWayLeverVisualizer
-          state_forward: switch-fwd
-          state_off: switch-off
-          state_reversed: switch-rev
+    - type: TwoWayLeverVisuals
+      state_forward: switch-fwd
+      state_off: switch-off
+      state_reversed: switch-rev
     - type: Damageable
       damageContainer: Inorganic
       damageModifierSet: Metallic
